Rename and added item support to tiles
Renamed IBuilding to IFeature and TileBehavior feature to special. The first was done to reflect that features should be modifiers of tiles, not inherent behaviors of tiles. Additionally, I think it is more correct for buildings to be contained over multiple tiles, so for now theoretically they will be groups of features. If a feature makes a tile unwalkable in any way, that tile will be a Special tile with its own characteristics (to be dealt with much later).