Adding 4 New Items
Wanted to push this directly onto main, but quickly saw that I have a few bugs to resolve with the new items, so I made a separate branch for them.
The bug seems to be that RangeChecker does not function correctly when given non-square hit maps. This is leading to attacks not originating from where they are supposed to and pivoting around weird points.
This commit includes the four new items, as well as minor modifications to the item systems to allow for naming items properly.
TODO: make changes to how items are read / work so that they don't need to go width-height-Y-X (desire X-Y)