Fixed bug with item reading
When creating the new weapons, I noticed that I had to put the holder's Y coord before their X coord in order for them to work. Today, I took the time to track down the bug, and begin the work of squashing it.
I tracked it down to an inconsistency between how I handle x and y (width and height) in GridManager and ItemReader + ItemPopulator. After finding the bug and resolving it, I decided to also change how the damageGrid gets stored so that 0,0 is the bottom left corner instead of the top left corner. This has ramifications that I will deal with in the next few commits.