Mana rework beginning
After starting the system previously, I decided to retool it to match the design I had in my head. For this reason, I spent some time working on a design document for the mana feature, which I am now beginning to implement.
The first change is that I want mana to be performant because mana is abundant, so I am changing it from a monobehavior style object to a struct.