Mana visualization progress!
Major progress with mana visualization!
I decided to go with a distributed visualization system which draws mana as orbs, and facilitates a transition between tiles.
As part of the system, I created a material (color) for every mana type.
Additionally, I modified UIClockManager and InitiativeManager to go from a 2-tick to a 3-tick system (which now includes ManaManager.

Open issues:
- ManaContainers should merge mana of the same type into one particle
- It seems like there is an issue where order of particles and orbs is not preserved properly. Need to investigate