Probabilistic system implemented
First draft of the new, probabilistic system has been implemented.
If a ManaContainer has received a new ManaParticle this phase, then it checks if it is now over the soft cap. If it is, then the ManaContainer asks politely if any point of mana wants to leave, and handles the leaving process if so.
TODO: ManaVisualizer is built around receiving every ManaParticle every phase, so it needs to be reworked.